+++ src/reducers/rootReducer.js
@@ -1,0 +1,19 @@
+const initState = {
+    user: {
+        "nome": "Arthur"
+    },
+    favorProfs: [
+        { id: 1, name: "Talita Cypriano", email: 'talita@gmail.com' },
+        { id: 2, name: "André Leme", email: 'andre@gmail.com' },
+        { id: 3, name: "Emílio Rodrigues", email: 'emilio@gmail.com' },
+        { id: 4, name: "Outro Professor", email: 'contato@gmail.com' },
+        { id: 5, name: "Outra Professora", email: 'contato@gmail.com' },
+        { id: 6, name: "Outra Professora", email: 'contato@gmail.com' },
+    ],
+}
+
+const rootReducer = (state = initState, action) => {
+    return state;
+}
+
+export default rootReducer